Swap the robot_client for individual nodes
